+++ NEWS.md
@@ -1,3 +1,16 @@
+# arf 0.2.4
+* Let verbose=FALSE silence (some) warnings
+
+# arf 0.2.3
+* Add impute() function for direct missing data imputation with ARF
+* Add one-line functions darf(), earf(), rarf()
+
+# arf 0.2.2
+* Faster and vectorized conditional sampling
+* Use min.bucket argument from ranger to avoid pruning if possible
+* Option to sample NAs in generated data if original data contains NAs
+* Stepsize in forge() to reduce memory usage
+* Option for local and global finite bounds
 
 # arf 0.2.0
 * Vectorized adversarial resampling